requirements.txt
# ════════════════════════════════════════════════════════════════════════════════
# Parker AI - Python Dependencies
# ════════════════════════════════════════════════════════════════════════════════

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0

# ════════════════════════════════════════════════════════════════════════════════
# LLM & AI
# ════════════════════════════════════════════════════════════════════════════════
langchain-core>=0.3.0
langchain-ollama>=0.2.0
langchain-groq>=0.2.0
langgraph>=0.2.0
langgraph-checkpoint-postgres>=1.0.0
langgraph-store-postgres>=1.0.0

# Embeddings
ollama>=0.1.0

# ════════════════════════════════════════════════════════════════════════════════
# Database
# ════════════════════════════════════════════════════════════════════════════════
psycopg[binary]>=3.1.0
# or psycopg2-binary>=2.9.0 for psycopg2

# ════════════════════════════════════════════════════════════════════════════════
# Voice
# ════════════════════════════════════════════════════════════════════════════════
faster-whisper>=1.0.0
pyttsx3>=2.90
sounddevice>=0.4.6
scipy>=1.10.0
numpy>=1.24.0

# Silero VAD for voice activity detection
# Install separately: pip install silero-vad
# Note: Requires torch, which is large. Optional dependency.

# ════════════════════════════════════════════════════════════════════════════════
# GUI
# ════════════════════════════════════════════════════════════════════════════════
PySide6>=6.6.0

# ════════════════════════════════════════════════════════════════════════════════
# API Server (optional)
# ════════════════════════════════════════════════════════════════════════════════
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0